Finished functionality for comments
Users can now delete their own comments.
Comments associated with an user will be deleted if that user is deleted.
Currently, admins can send comment to other users (if they happen to be in a group with them). Also, users can comment on their group members' pages regardless of the day.